backend/requirements.txt
fastapi
uvicorn
python-multipart
SQLAlchemy
pydantic
psycopg2
Pillow==2.6.0
